feat: alterando tipo da média da resenha para Double e documentando endpoints com Swagger
Este commit altera o tipo da média da resenha para `Double` e realiza ajustes no `ResenhaViewDTO` para refletir essa mudança. Além disso, documenta com Swagger os controllers `RelacionamentoSeguidoresController`, `FavoritoController` e `AvaliacaoController`, bem como seus respectivos DTOs. As mudanças visam melhorar a precisão dos dados, a clareza da documentação e a consistência da API.

Mudanças Principais:

- Alteração do tipo da média da resenha para `Double`:
  - O campo `media_resenha` na entidade `Resenha` foi alterado de `Integer` para `Double` para permitir valores decimais.
  - Ajustes no `ResenhaViewDTO` para refletir a mudança do tipo da média.

- Documentação com Swagger:
  - RelacionamentoSeguidoresController:
    - Documentados endpoints.
    - Adicionadas descrições detalhadas e exemplos para cada endpoint.
  - FavoritoController:
   - Documentados endpoints.
   - Adicionadas descrições e exemplos para os DTOs `FavoritoRequestDTO` e `FavoritoResponseDTO`.
  - AvaliacaoController:
    - Documentados endpoints.
    - Adicionadas descrições e exemplos para o DTO `AvaliacaoDTO`.

+++ src/main/java/prati/projeto/redeSocial/rest/dto/ResenhaViewDTO.java
@@ -19,6 +19,6 @@ public class ResenhaViewDTO {
     private String dataEdicao;
     private Double nota;
     private boolean spoiler;
-    private Integer media;
+    private Double media;
     private List<AvaliacaoDTO> avaliacoes;
 }
